Repair tests create, replace, getall When converted public static into public transUtils method there was an error that told me that i should start mocking all the methods i use.

+++ src/main/java/com/github/WeeiaEduTeam/InfinityFinanceAPI/util/Util.java
@@ -1,7 +1,7 @@
 package com.github.WeeiaEduTeam.InfinityFinanceAPI.util;
 
 public class Util {
-    public static boolean isPositive(int number) {
+    public static boolean isPositive(long number) {
         return number > 0 ? Boolean.TRUE : Boolean.FALSE;
     }
 }
